Move Pin over to /Discord instead of /School
